Removes PendingItem.
This is a huge change. Now the torrents package queries the show for all
the relevant information. I found out that losing the coupling between
the PendingItem and Show was very annoying.

Also significant is the introduction of Snippets. Some shows have long
name while ppl online use an abbreviation of it. Consider Snippets as
alternatives for a search query. Snippets are stored and read, this
prevents doing the same query again and again like was the case with the
old QueryNames.

There is still a ton to do in kickass.go. Season and episode search are
more or less copy pastes...