Significantly simplified snippet selection.
God this annoyed my for the longest time.
The behaviour changed a bit;
It used to be the case that, when no current snippet was ever stored,
the code would generate _all_ differnt alternative snippets. It would
then select the best and would use that for ever.
That last part sucked.

Now it will explore 10% of the time.